+++ Changelog.md
@@ -15,6 +15,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Prefabs for bullets (26/08/2025)
  - Player bullet shoot (26/08/2025)
  - Song start script (26/08/2025)
+ - Bullets pool (28/08/2025)
+ - Base enemy script (28/08/2025)
+ - Final player sprite (28/08/2025)
+
+### Modified
+
+ - Sound manager script (28/08/2025)
 
 ## [0.0.0] - 25/08/2025